+++ CHANGELOG.md
@@ -2,6 +2,10 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2024.01.22`.
+ * Upgraded stable Dart analysis SDK to `3.2.5`.
+ * Upgraded stable Flutter analysis SDK to `3.16.8`.
+ * Upgraded preview Flutter analysis SDK to `3.19.0-0.2.pre`.
 
 ## `20240118t104800-all`
  * Upgraded dependencies including `markdown: ^7.2.0`.